move mirror controller to new world
