+++ subiquity/common/apidef.py
@@ -41,6 +41,7 @@ class API:
     locale = simple_endpoint(str)
     keyboard = simple_endpoint(KeyboardSetting)
     proxy = simple_endpoint(str)
+    mirror = simple_endpoint(str)
 
     class meta:
         class status:
